requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
python-magic>=0.4.27

# ML and embeddings
openai>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0

# Data handling
pandas>=2.1.0
pydantic>=2.4.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
loguru>=0.7.2

# Email processing
email-validator>=2.1.0
